Kernel compiler warnings cleanup - modem
    [Defect / Enhancement / New Feature]
    MobC00184944

    [Problem]
    Remove warnings as listed in the attached modem.txt file

    [Solution]
    Remove compiler warnings

    [Reviewers]
    Jin Guo, Mike Gallant

    [Testing]
        [Test Cases]
            Should not affect AP build, all warnings in modem.txt should be gone

        [Label/CL# Synced To]
            sdb-rhea-android-ics

        [Platform]
            RheaRay

    [BROADCOM INTERNAL NOTES]

+++ drivers/net/brcm_netconsole.c
@@ -401,7 +401,7 @@ static ssize_t store_enabled(struct brcm_netconsole_target *nt,
 	if (enabled < 0)
 		return enabled;
 
-	pr_info("%s enabled=%d\n", __func__, enabled);
+	pr_info("%s enabled=%d\n", __func__, (int)enabled);
 
 	if (enabled) {	/* 1 */
